moved function to r_packet.py
